Renaming of resources file to constant
